selenium-ui-automation/requirements.txt
selenium==4.36.0
pytest==8.3.3
webdriver-manager==4.0.2
allure-pytest==2.13.5


